screenshot

+++ app/mathematical_calculator.c
@@ -11,8 +11,8 @@ void discriminant(double a, double b, double c, double* x1, double* x2, int* has
     if (a != 0) {
         d = b * b - 4 * a * c;
         if (d > 0) {
-            *x1 = (double)((-b - sqrt(d)) / (2 * a));
-            *x2 = (double)((-b + sqrt(d)) / (2 * a));
+            *x1 = (-b - sqrt(d)) / (2 * a);
+            *x2 = ((-b + sqrt(d)) / (2 * a);
         } else if (d == 0) {
             *x1 = *x2 = -b / (2 * a);
         } else {
